+++ python/llm/src/bigdl/llm/vllm/model_executor/__init__.py
@@ -14,7 +14,7 @@
 # limitations under the License.
 #
 
-from bigdl.llm.vllm.models.model_loader import get_model
+from bigdl.llm.vllm.model_executor.model_loader import get_model
 
 __all__ = [
     "get_model",
